added Linker

+++ src/List.c
@@ -101,7 +101,7 @@ List* listXOR(List* l1, List* l2){
 
   le = l2->head;
 
-  for (size_t i = 0; i < l2->size; i++) {
+  for (int i = 0; i < l2->size; i++) {
     if (searchInList(l1, le->data) == -1) {
       insert(newList, le->data);
     }
@@ -115,7 +115,7 @@ List* listComplement(List* l1, List* l2){
   List* newList = initList();
   ListElement* le = l1->head;
 
-  for (size_t i = 0; i < l1->size; i++) {
+  for (int i = 0; i < l1->size; i++) {
     if (searchInList(l2, le->data) == -1) {
       insert(newList, le->data);
     }
